# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  masterarea.CreateRequest:
    properties:
      description:
        maxLength: 200
        type: string
      level_area_id:
        maxLength: 20
        type: string
      location:
        maxLength: 100
        type: string
      name:
        maxLength: 50
        type: string
      parent_id:
        minimum: 1
        type: integer
      status:
        enum:
        - active
        - inactive
        type: string
    required:
    - description
    - level_area_id
    - location
    - name
    type: object
  masterarea.UpdateRequest:
    properties:
      description:
        maxLength: 200
        type: string
      level_area_id:
        maxLength: 20
        type: string
      location:
        maxLength: 100
        type: string
      name:
        maxLength: 50
        type: string
      parent_id:
        minimum: 1
        type: integer
      status:
        enum:
        - active
        - inactive
        type: string
    required:
    - description
    - level_area_id
    - location
    - name
    type: object
  masterdonationcategory.CreateRequest:
    properties:
      name:
        maxLength: 100
        type: string
      status:
        enum:
        - active
        - inactive
        type: string
    required:
    - name
    type: object
  masterdonationcategory.UpdateRequest:
    properties:
      name:
        maxLength: 100
        type: string
      status:
        enum:
        - active
        - inactive
        type: string
    required:
    - name
    type: object
  masterdonatur.CreateRequest:
    properties:
      area:
        maxLength: 100
        type: string
      id_donatur:
        maxLength: 50
        type: string
      id_group_donatur:
        type: integer
      id_tzu_chi_app:
        maxLength: 50
        type: string
      id_vis_relawan:
        maxLength: 50
        type: string
      status:
        enum:
        - active
        - inactive
        type: string
      telepon:
        maxLength: 20
        type: string
    required:
    - area
    - id_donatur
    - telepon
    type: object
  masterdonatur.UpdateRequest:
    properties:
      area:
        maxLength: 100
        type: string
      id_donatur:
        maxLength: 50
        type: string
      id_group_donatur:
        type: integer
      id_tzu_chi_app:
        maxLength: 50
        type: string
      id_vis_relawan:
        maxLength: 50
        type: string
      status:
        enum:
        - active
        - inactive
        type: string
      telepon:
        maxLength: 20
        type: string
    required:
    - area
    - id_donatur
    - telepon
    type: object
...
info:
  contact: {}
  description: API documentation
  title: Social BE API
  version: "1.0"
paths:
...
  /master/donation-categories:
    get:
      description: Get list of master donation category
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get all master donation category
      tags:
      - master-donation-category
    post:
      consumes:
      - application/json
      description: Create new master donation category
      parameters:
      - description: Create master donation category
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/masterdonationcategory.CreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create master donation category
      tags:
      - master-donation-category
  /master/donation-categories/{id}:
    delete:
      description: Soft delete master donation category
      parameters:
      - description: Master Donation Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete master donation category
      tags:
      - master-donation-category
    get:
      description: Get detail master donation category
      parameters:
      - description: Master Donation Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get master donation category by ID
      tags:
      - master-donation-category
    put:
      consumes:
      - application/json
      description: Update existing master donation category
      parameters:
      - description: Master Donation Category ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update master donation category
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/masterdonationcategory.UpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update master donation category
      tags:
      - master-donation-category
...
  /master/donaturs:
    get:
      description: Get list of master donatur
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get all master donatur
      tags:
      - master-donatur
    post:
      consumes:
      - application/json
      description: Create new master donatur
      parameters:
      - description: Create master donatur
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/masterdonatur.CreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create master donatur
      tags:
      - master-donatur
  /master/donaturs/{id}:
    delete:
      description: Soft delete master donatur
      parameters:
      - description: Master Donatur ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete master donatur
      tags:
      - master-donatur
    get:
      description: Get detail master donatur
      parameters:
      - description: Master Donatur ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get master donatur by ID
      tags:
      - master-donatur
    put:
      consumes:
      - application/json
      description: Update existing master donatur
      parameters:
      - description: Master Donatur ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update master donatur
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/masterdonatur.UpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update master donatur
      tags:
      - master-donatur
...
  /master/master-areas:
    get:
      description: Get list of master area
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get all master area
      tags:
      - master-area
    post:
      consumes:
      - application/json
      description: Create new master area
      parameters:
      - description: Create master area
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/masterarea.CreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create master area
      tags:
      - master-area
  /master/master-areas/{id}:
    delete:
      description: Soft delete master area
      parameters:
      - description: Master Area ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete master area
      tags:
      - master-area
    get:
      description: Get detail master area
      parameters:
      - description: Master Area ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get master area by ID
      tags:
      - master-area
    put:
      consumes:
      - application/json
      description: Update existing master area
      parameters:
      - description: Master Area ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update master area
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/masterarea.UpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update master area
      tags:
      - master-area
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"